ref: Rename validate.ts to options.ts to be more descriptive
